Split agents.py back into per-agent modules
Remove the monolithic agents/agents.py and define each agent in its own
module again, restoring the per-file layout while keeping main's catalog,
CoderAgent, and dedicated coder model:

- agents/common.py: shared setup (make_llm, make_coder_llm, agent_tools,
  MODEL/CODER_MODEL, hitl, litellm config).
- one module per agent (hypotheses, research, planner, medical, coder) with
  the real LlmAgent definition; task_execution_agent.py holds the full
  tool-discovery + FEDOT pipeline.
- orchestrator_agent.py owns the catalog wiring (_AGENT_INSTANCES,
  build_orchestrator_instruction, pre_action_critique) and recursive tracing.
- catalog.py / docs/a2a.md: point to the per-module layout.

Verified: full package + A2A_MODE import, all 6 sub-agent servers + orchestrator
boot, live orchestrator->HypothesesAgent delegation over A2A completes, pyflakes
clean.
