+++ CoScientist/agents/orchestrator_agent.py
@@ -1,9 +1,67 @@
-"""Re-export shim for the in-process OrchestratorAgent.
+"""OrchestratorAgent — top-level coordinator (in-process ADK mode).
 
-The agent is defined in :mod:`CoScientist.agents.agents` (single source of
-truth, driven by the agent catalog). For the A2A orchestrator (sub-agents as
-remote services) see :mod:`CoScientist.a2a.orchestrator`.
+The roster is driven by the agent catalog (:mod:`CoScientist.agents.catalog`) —
+the single source of truth for which agents are enabled, their prompt
+descriptions, and their order. We map each catalog name to its LlmAgent
+instance and attach the enabled ones as tools.
+
+For A2A mode (each sub-agent as a remote HTTP service) see
+:mod:`CoScientist.a2a.orchestrator`.
 """
-from CoScientist.agents.agents import orchestrator_agent
+from google.adk.agents.llm_agent import LlmAgent
+from google.adk.tools.agent_tool import AgentTool
+
+from CoScientist.agents import catalog
+from CoScientist.agents.coder_agent import coder_agent
+from CoScientist.agents.common import agent_tools, make_llm
+# post_action_critique is intentionally not wired (after_tool_callback disabled below).
+from CoScientist.agents.critic_agent import pre_action_critique
+from CoScientist.agents.hypotheses_agent import hypotheses_agent
+from CoScientist.agents.med_callbacks import before_model_modifier as med_before_model
+from CoScientist.agents.medical_agent import medical_agent
+from CoScientist.agents.planner_agent import planner_agent
+from CoScientist.agents.prompts import build_orchestrator_instruction
+from CoScientist.agents.research_agent import research_agent
+from CoScientist.agents.task_execution_agent import task_execution_agent
+from CoScientist.logging import multi_agent_tracer
+from opik.integrations.adk import track_adk_agent_recursive
+
+# Catalog name -> the LlmAgent instance it refers to.
+_AGENT_INSTANCES = {
+    "PlannerAgent": planner_agent,
+    "HypothesesAgent": hypotheses_agent,
+    "ResearchAgent": research_agent,
+    "TaskExecutorAgent": task_execution_agent,
+    "MedicalAgent": medical_agent,
+    "CoderAgent": coder_agent,
+}
+
+
+def _resolve_agent(name: str):
+    inst = _AGENT_INSTANCES.get(name)
+    if inst is None:
+        raise ValueError(
+            f"Catalog agent {name!r} has no instance in _AGENT_INSTANCES "
+            "(orchestrator_agent.py). Add it there or fix the catalog name."
+        )
+    return inst
+
+
+_orchestrator_subagents = [
+    AgentTool(agent=_resolve_agent(spec.name)) for spec in catalog.enabled_agents()
+]
+
+orchestrator_agent = LlmAgent(
+    name="OrchestratorAgent",
+    model=make_llm(),
+    instruction=build_orchestrator_instruction(),
+    description="Main Orchestrator Agent",
+    before_model_callback=med_before_model,
+    after_model_callback=pre_action_critique,
+    # after_tool_callback=post_action_critique,
+    tools=agent_tools(_orchestrator_subagents, hitl=False),
+)
+
+track_adk_agent_recursive(orchestrator_agent, multi_agent_tracer)
 
 __all__ = ["orchestrator_agent"]
